feat: 버그 수정, Postgres 로 고정
