+++ 09-spring/06-exposed-spring-web-multitenant/src/main/kotlin/exposed/multitenant/springweb/ExposedMultitenantApplication.kt
@@ -8,8 +8,10 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.WebApplicationType
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
+import org.springframework.context.annotation.EnableAspectJAutoProxy
 
 @SpringBootApplication
+@EnableAspectJAutoProxy
 class ExposedMultitenantApplication {
 
     companion object: KLogging()
