+++ 09-spring/06-exposed-spring-web-multitenant/src/main/kotlin/exposed/multitenant/springweb/tenant/TenantService.kt
@@ -2,7 +2,6 @@ package exposed.multitenant.springweb.tenant
 
 import io.bluetape4k.logging.KLogging
 import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
 
 @Service
 class TenantService(
@@ -10,7 +9,6 @@ class TenantService(
 ) {
     companion object: KLogging()
 
-    @Transactional
     fun createTenantSchemas() {
         Tenants.Tenant.entries.forEach { tenant ->
             TenantContext.setCurrentTenant(tenant)
