+++ 09-spring/06-exposed-spring-web-multitenant/src/main/kotlin/exposed/multitenant/springweb/tenant/TransactionSchemaAspect.kt
@@ -1,27 +1,31 @@
 package exposed.multitenant.springweb.tenant
 
-import org.aspectj.lang.ProceedingJoinPoint
+import io.bluetape4k.logging.KLogging
+import io.bluetape4k.logging.info
 import org.aspectj.lang.annotation.Aspect
 import org.aspectj.lang.annotation.Before
 import org.jetbrains.exposed.sql.SchemaUtils
+import org.springframework.core.Ordered
+import org.springframework.core.annotation.Order
 import org.springframework.stereotype.Component
-import org.springframework.transaction.annotation.Transactional
 
 @Aspect
 @Component
+@Order(Ordered.HIGHEST_PRECEDENCE + 1)
 class TransactionSchemaAspect {
 
-    @Aspect
-    @Component
-    class TransactionSchemaAspect {
+    companion object: KLogging()
 
-        @Before("@annotation(org.springframework.transaction.annotation.Transactional)")
-        fun setSchemaForTransaction(joinPoint: ProceedingJoinPoint, transactional: Transactional): Any? {
-            runCatching {
-                val schema = TenantContext.getCurrentTenantSchema()
-                SchemaUtils.setSchema(schema)
-            }
-            return joinPoint.proceed()
+    /**
+     * `@Transactional` 이 적용된 메소드에 대해 Multitenncy 를 지원하기 위한 Schema 설정을 수행합니다.
+     */
+    @Before("@within(org.springframework.transaction.annotation.Transactional)")
+    fun setSchemaForTransaction() {
+        runCatching {
+            val schema = TenantContext.getCurrentTenantSchema()
+            log.info { "Use schema=$schema" }
+            SchemaUtils.createSchema(schema)
+            SchemaUtils.setSchema(schema)
         }
     }
 }
