+++ 09-spring/06-exposed-spring-web-multitenant/src/main/resources/application.yml
@@ -18,9 +18,6 @@ spring:
     application:
         name: Spring MVC with Exposed
 
-    profiles:
-        default: "postgres"    # postgres | h2 를 선택할 수 있습니다.
-
     threads:
         virtual:
             enabled: true   # Virtual Thread 사용 여부
